Turmult.Gateway.Types

Description

Documentation

module Turmult.Gateway.Types.Messages

data GatewayInfo

Data constructors to discover Gateway connection info - https://discord.com/developers/docs/topics/gateway#get-gateway-bot

url :: Lens' GatewayInfo Text

shards :: Lens' GatewayInfo Integer

session_start_limit :: Lens' GatewayInfo SessionStartLimit

data SessionStartLimit

Constructors

SessionStartLimit

Fields

_total :: Integer

The total number of session starts the current user is allowed

_remaining :: Integer

The remaining number of session starts the current user is allowed

_reset_after :: Integer

The number of milliseconds after which the limit resets

_max_concurrency :: Integer

The number of identify requests allowed per 5 seconds

total :: Lens' SessionStartLimit Integer

remaining :: Lens' SessionStartLimit Integer

max_concurrency :: Lens' SessionStartLimit Integer

reset_after :: Lens' SessionStartLimit Integer

data GatewayParams

version :: Lens' GatewayParams Integer

encoding :: Lens' GatewayParams GatewayEncoding

compress :: Lens' GatewayParams (Maybe GatewayCompression)

data GatewayEncoding

The Discord Gateway supports both JSON and ETF encodings. - We, however, only support JSON at the moment.

Constructors

JSON

ETF

data GatewayCompression

The Discord Gateway supports ZLIB compression on payloads - Unfortunately, we do not.

Constructors

ZLIB_STREAM