Turmult.Gateway.Types.Messages

Description

Documentation

Gateway Commands

data Payload

Payload format for all messages sent/received over Gateway socket - https://discord.com/developers/docs/topics/gateway#payloads

Constructors

Payload

Fields

_payload_op :: Integer

opcode for the payload

_payload_d :: Maybe Value

event data

_payload_s :: Maybe Integer

sequence number, used for resuming sessions and heartbeats

_payload_t :: Maybe Text

the event name for this payload

payload_op :: Lens' Payload Integer

payload_d :: Lens' Payload (Maybe Value)

payload_s :: Lens' Payload (Maybe Integer)

payload_t :: Lens' Payload (Maybe Text)

data Identify

Opcode 2 Identify: Send - Starts a new session during the initial handshake.

Constructors

Identify

Fields

_identify_token :: Text

authentication token

_identify_properties :: Properties

connection properties

_identify_compress :: Maybe Bool

whether this connection supports compression of packets - Default: False

_identify_large_threshold :: Maybe Int

value between 50 and 250, total number of members where the gateway will stop sending offline members in the guild member list

_identify_shard :: Maybe Shard

used for Guild Sharding

_identify_presence :: Maybe Presence

presence structure for initial presence information

_identify_guild_subscriptions :: Maybe Bool

enables dispatching of guild subscription events (presence and typing events)

_identify_intents :: GatewayIntents

the Gateway Intents you wish to receive

type Snowflake = Text

Discord utilizes Twitter's snowflake format for uniquely identifiable - descriptors (IDs). These IDs are guaranteed to be unique across all of - Discord, except in some unique scenarios in which child objects share their - parent's ID.

type Heartbeat = Maybe Int

Opcode 0 Dispatch: Receive - An event was dispatched. - This is encoded as one of the event payloads defined further down

Opcode 1 Heartbeat: Send/Receive - Fired periodically by the client to keep the connection alive.

data Presence

Opcode 3 Presence Update: Send - Update the client's presence.

Constructors

Presence

Fields

_presence_since :: Maybe Int

unix time (in milliseconds) of when the client went idle, or null if the client is not idle

_presence_activities :: [Activity]

null, or the user's activities

_presence_status :: StatusType

the user's new status

_presence_afk :: Bool

whether or not the client is afk

data VoiceStateUpdate

Opcode 4 Voice State Update: Send - Used to join/leave or move between voice channels.

Constructors

VoiceStateUpdate

Fields

_voice_state_update_token :: Text

voice connection token

_voice_state_update_guild_id :: Snowflake

the guild this voice server update is for

_voice_state_update_endpoint :: Text

the voice server host

data Resume

Opcode 6 Resume: Send - Resume a previous session that was disconnected.

Constructors

Resume

Fields

_resume_token :: Text

session token

_resume_session_id :: Text

session id

_resume_seq :: Int

last sequence number received

type Reconnect = Bool

Opcode 7 Reconnect: Receive - You should attempt to reconnect and resume immediately. - indicates whether the session may be resumable.

data RequestGuildMembers

Opcode 8 Request Guild Members: Send - Request information about offline guild members in a large guild.

Constructors

RequestGuildMembers

Fields

_request_guild_members_guild_id :: Snowflake

id of the guild to get members for

_request_guild_members_query :: Maybe Text

Text that username starts with, or an empty Text to return all members

_request_guild_members_limit :: Int

maximum number of members to send matching the query; a limit of 0 can be used with an empty Text query to return all members

_request_guild_members_presences :: Maybe Bool

used to specify if we want the presences of the matched members

_request_guild_members_user_ids :: Maybe [Snowflake]

used to specify which users you wish to fetch

_request_guild_members_nonce :: Maybe Text

nonce to identify the Guild Members Chunk response

type InvalidSession = Bool

Opcode 9 Invalid Session: Receive - The session has been invalidated. You should reconnect and identify/resume accordingly. - indicates whether the session may be resumable.

newtype Hello

Opcode 10 Hello: Receive - Sent immediately after connecting, contains the heartbeat_interval to use.

Constructors

Hello

Fields

_hello_heartbeat_interval :: Int

the interval (in milliseconds) the client should heartbeat with

data Properties

Connection Properties

Constructors

Properties

Fields

_properties_os :: Text

your operating system

_properties_browser :: Text

your library name

_properties_device :: Text

your library name

data Shard

Shard Identifier

Constructors

Shard

Fields

_shard_id :: Int

Shard ID

_shard_num :: Int

Number of shards

data StatusType

Constructors

Online
Online

DND
Do Not Disturb

Idle
AFK

Invisible
Invisible and shown as offline

Offline
Offline

data GatewayIntents

Bit shifting object representing gateway intents that can be subscribed to

Constructors

Intents

Fields

_gateway_intents_guilds :: Bool

_gateway_intents_guild_members :: Bool

_gateway_intents_guild_bans :: Bool

_gateway_intents_guild_emojis :: Bool

_gateway_intents_guild_integrations :: Bool

_gateway_intents_guild_webhooks :: Bool

_gateway_intents_guild_invites :: Bool

_gateway_intents_guild_voice_status :: Bool

_gateway_intents_guild_presences :: Bool

_gateway_intents_guild_messages :: Bool

_gateway_intents_guild_message_reactions :: Bool

_gateway_intents_guild_message_typing :: Bool

_gateway_intents_direct_messages :: Bool

_gateway_intents_direct_message_reactions :: Bool

_gateway_intents_direct_message_typing :: Bool

data Ready

Constructors

Ready

Fields

_ready_v :: Int

_ready_user :: User

_ready_private_channels :: [Channel]

_ready_guilds :: [Guild]

_ready_session_id :: Text

_ready_shard :: Maybe Shard

_ready_application :: Object