Turmult.Types.Data

Description

Type Classes

class ToJSON a => Sendable a

class FromJSON a => Receivable a

Resources

General

type Snowflake = Text

Discord utilizes Twitter's snowflake format for uniquely identifiable - descriptors (IDs). These IDs are guaranteed to be unique across all of - Discord, except in some unique scenarios in which child objects share their - parent's ID.

Audit Log

data AuditLog

Audit Logs - Whenever an admin action is performed on the API, an entry is added to the - respective guild's audit log. You can specify the reason by attaching the - `X-Audit-Log-Reason` request header. This header supports url encoded utf8 - characters.

data AuditLogEntry

Audit Log Entry Object

data AuditLogChange

Audit Log Change Object If new_value is not present in the change object, while old_value is, that means the property that was changed has been reset, or set to null

data AuditLogEvent

Audit Log Events

Constructors

GUILD_UPDATE

CHANNEL_CREATE

CHANNEL_UPDATE

CHANNEL_DELETE

CHANNEL_OVERWRITE_CREATE

CHANNEL_OVERWRITE_UPDATE

CHANNEL_OVERWRITE_DELETE

MEMBER_KICK

MEMBER_PRUNE

MEMBER_BAN_ADD

MEMBER_BAN_REMOVE

MEMBER_UPDATE

MEMBER_ROLE_UPDATE

MEMBER_MOVE

MEMBER_DISCONNECT

BOT_ADD

ROLE_CREATE

ROLE_UPDATE

ROLE_DELETE

INVITE_CREATE

INVITE_UPDATE

INVITE_DELETE

WEBHOOK_CREATE

WEBHOOK_UPDATE

WEBHOOK_DELETE

EMOJI_CREATE

EMOJI_UPDATE

EMOJI_DELETE

MESSAGE_DELETE

MESSAGE_BULK_DELETE

MESSAGE_PIN

MESSAGE_UNPIN

INTEGRATION_CREATE

INTEGRATION_UPDATE

INTEGRATION_DELETE

data AuditLogChangeType

Constructors

AuditLogChangedGuild

AuditLogChangedChannel

AuditLogChangedRole

AuditLogChangedInvite

AuditLogChangedUser

AuditLogChangedAny

AuditLogChangedIntegration

AuditLogChangedVoiceChannel

data AuditLogChangeValue

Constructors

AuditLogChangeName (Maybe Text) (Maybe Text)
Change Type: guild - name changed

AuditLogChangeDescription (Maybe Text) (Maybe Text)
Change Type: guild - description changed

AuditLogChangeIconHash (Maybe Text) (Maybe Text)
Change Type: guild - icon changed

AuditLogChangeSplashHash (Maybe Text) (Maybe Text)
Change Type: guild - invite splash page artwork changed

AuditLogChangeDiscoverySplashHash (Maybe Text) (Maybe Text)
Change Type: guild - discovery splash changed

AuditLogChangeBannerHash (Maybe Text) (Maybe Text)
Change Type: guild - guild banner changed

AuditLogChangeOwnerId (Maybe Snowflake) (Maybe Snowflake)
Change Type: guild - owner changed

AuditLogChangeRegion (Maybe Text) (Maybe Text)
Change Type: guild - region changed

AuditLogChangePreferredLocale (Maybe Text) (Maybe Text)
Change Type: guild - preferred locale changed

AuditLogChangeAfkChannelId (Maybe Snowflake) (Maybe Snowflake)
Change Type: guild - afk channel changed

AuditLogChangeAfkTimeout (Maybe Int) (Maybe Int)
Change Type: guild - afk timeout duration changed

AuditLogChangeRulesChannelId (Maybe Snowflake) (Maybe Snowflake)
Change Type: guild - id of the rules channel changed

AuditLogChangePublicUpdatesChannelId (Maybe Snowflake) (Maybe Snowflake)
Change Type: guild - id of the public updates channel changed

AuditLogChangeMfaLevel (Maybe Int) (Maybe Int)
Change Type: guild - two-factor auth requirement changed

AuditLogChangeVerificationLevel (Maybe Int) (Maybe Int)
Change Type: guild - required verification level changed

AuditLogChangeExplicitContentFilter (Maybe Int) (Maybe Int)
Change Type: guild - change in whose messages are scanned and deleted for explicit content in the server

AuditLogChangeDefaultMessageNotifications (Maybe Int) (Maybe Int)
Change Type: guild - default message notification level changed

AuditLogChangeVanityUrlCode (Maybe Text) (Maybe Text)
Change Type: guild - guild invite vanity url changed

AuditLogChangeAdd (Maybe [Role]) (Maybe [Role])
Change Type: guild - new role added

AuditLogChangeRemove (Maybe [Role]) (Maybe [Role])
Change Type: guild - role removed

AuditLogChangePruneDeleteDays (Maybe Int) (Maybe Int)
Change Type: guild - change in number of days after which inactive and role-unassigned members are kicked

AuditLogChangeWidgetEnabled (Maybe Bool) (Maybe Bool)
Change Type: guild - server widget enabled/disable

AuditLogChangeWidgetChannelId (Maybe Snowflake) (Maybe Snowflake)
Change Type: guild - channel id of the server widget changed

AuditLogChangeSystemChannelId (Maybe Snowflake) (Maybe Snowflake)
Change Type: guild - id of the system channel changed

AuditLogChangePosition (Maybe Int) (Maybe Int)
Change Type: channel - text or voice channel position changed

AuditLogChangeTopic (Maybe Text) (Maybe Text)
Change Type: channel - text channel topic changed

AuditLogChangeBitrate (Maybe Int) (Maybe Int)
Change Type: channel - voice channel bitrate changed

AuditLogChangePermissionOverwrites (Maybe [Channel]) (Maybe [Channel])
Change Type: channel - permissions on a channel changed

AuditLogChangeNsfw (Maybe Bool) (Maybe Bool)
Change Type: channel - channel nsfw restriction changed

AuditLogChangeApplicationId (Maybe Snowflake) (Maybe Snowflake)
Change Type: channel - application id of the added or removed webhook or bot

AuditLogChangeRateLimitPerUser (Maybe Int) (Maybe Int)
Change Type: channel - amount of seconds a user has to wait before sending another message changed

AuditLogChangePermissions (Maybe Text) (Maybe Text)
Change Type: role - permissions for a role changed

AuditLogChangeColor (Maybe Int) (Maybe Int)
Change Type: role - role color changed

AuditLogChangeHoist (Maybe Bool) (Maybe Bool)
Change Type: role - role is now displayed/no longer displayed separate from online users

AuditLogChangeMentionable (Maybe Bool) (Maybe Bool)
Change Type: role - role is now mentionable/unmentionable

AuditLogChangeAllow (Maybe Text) (Maybe Text)
Change Type: role - a permission on a text or voice channel was allowed for a role

AuditLogChangeDeny (Maybe Text) (Maybe Text)
Change Type: role - a permission on a text or voice channel was denied for a role

AuditLogChangeCode (Maybe Text) (Maybe Text)
Change Type: invite - invite code changed

AuditLogChangeChannelId (Maybe Snowflake) (Maybe Snowflake)
Change Type: invite - channel for invite code changed

AuditLogChangeInviterId (Maybe Snowflake) (Maybe Snowflake)
Change Type: invite - person who created invite code changed

AuditLogChangeMaxUses (Maybe Int) (Maybe Int)
Change Type: invite - change to max number of times invite code can be used

AuditLogChangeUses (Maybe Int) (Maybe Int)
Change Type: invite - number of times invite code used changed

AuditLogChangeMaxAge (Maybe Int) (Maybe Int)
Change Type: invite - how long invite code lasts changed

AuditLogChangeTemporary (Maybe Bool) (Maybe Bool)
Change Type: invite - invite code is temporary/never expires

AuditLogChangeDeaf (Maybe Bool) (Maybe Bool)
Change Type: user - user server deafened/undeafened

AuditLogChangeMute (Maybe Bool) (Maybe Bool)
Change Type: user - user server muted/unmuted

AuditLogChangeNick (Maybe Text) (Maybe Text)
Change Type: user - user nickname changed

AuditLogChangeAvatarHash (Maybe Text) (Maybe Text)
Change Type: user - user avatar changed

AuditLogChangeId (Maybe Snowflake) (Maybe Snowflake)
Change Type: any - the id of the changed entity - sometimes used in conjunction with other keys

AuditLogChangeType (Maybe Text) (Maybe Text)
Change Type: any - type of entity created

AuditLogChangeEnableEmoticons (Maybe Bool) (Maybe Bool)
Change Type: integration - integration emoticons enabled/disabled

AuditLogChangeExpireBehavior (Maybe Int) (Maybe Int)
Change Type: integration - integration expiring subscriber behavior changed

AuditLogChangeExpireGracePeriod (Maybe Int) (Maybe Int)
Change Type: integration - integration expire grace period changed

AuditLogChangeUserLimit (Maybe Int) (Maybe Int)
Change Type: voice channel - new user limit in a voice channel

data AuditLogInfo

Optional Audit Entry Info

Unsorted

data Activity

data ActivityType

data Timestamps

data ActivityEmoji

data ActivityParty

data PartySize

data ActivityAssets

data ActivitySecrets

data Channel

data ClientStatus

data Emoji

data Guild

data GuildMember

data Invite

data Overwrite

data OverwriteType

data PresenceUpdate

data Role

data RoleTag

data Template

data User

data UserFlags

data VoiceState

data Webhook

data WebhookType

data WelcomeScreen

data WelcomeScreenChannel